Simplify Food Registration with a registerFood method on ModItems
This commit simplifies the food registration process in ModItems by introducing a new method called registerFood. This method takes in the name of the item and its corresponding food properties, and registers the item with the specified properties. By using this method, we can now easily register multiple food items without duplicating code.